fix: icon is not displayed after package installed (#57)
When we install the package which binary packed in /opt, the trigger
will link the desktop file and binary to /usr. But it links desktop
file first and AM can not find the program of the package in any path
of $PATH at that moment. Then the signal which named ItemChanged is not
emited by AM and launcher will not display the icon.

This is a quick fix, wait 1000ms until the binary link is complete.

Log:

Co-authored-by: kamiyadm <chenhuixing@deepin.org>