fix: support ExecSearchPath to prevent systemd from not finding binaries
In NixOS, maintainer used a patch to modify the DEFAULT_PATH macro of systemd.
However, systemd can only find its own binaries under that path.
At this point, if you take the relative path of the binary as a parameter to StartTransientUnit,
systemd will use DEFAULT_PATH to look for the binary, which will cause systemd to not find it.

There may be other Linux distributions that change this path,
so add ExecSearchPath to prevent systemd from not finding the binary.

Signed-off-by: ComixHe <heyuming@deepin.org>